+++ CHANGELOG.md
@@ -671,3 +671,4 @@ This is a fast forward to v1.6.3 of daviscook477's fork with a few additional ch
 * Fix crash if mods add unlocks badly (kiooeht)
 
 #### dev ####
+* Fix hi-res relic art not being used on mod relics that have it (kiooeht)